Add ResultDto

+++ University_CQRS/Handlers/UnregisterCommandHandlerPostProcessor.cs
@@ -1,13 +1,10 @@
-﻿
-using MediatR;
-using MediatR.Pipeline;
+﻿using MediatR.Pipeline;
 using University_CQRS.Commands;
-
 using University_CQRS.Persistance.Repositories;
 
 namespace University_CQRS.Handlers
 {
-    public class UnregisterCommandHandlerPostProcessor : IRequestPostProcessor<UnregisterCommand, Unit>
+    public class UnregisterCommandHandlerPostProcessor : IRequestPostProcessor<UnregisterCommand, ResultDto>
     {
         private readonly StudentReadRepository _studentReadRepository;
 
@@ -16,11 +13,14 @@ namespace University_CQRS.Handlers
             _studentReadRepository = studentReadRepository;
         }
 
-        public Task Process(UnregisterCommand request, Unit response, CancellationToken cancellationToken)
+        public Task Process(UnregisterCommand request, ResultDto response, CancellationToken cancellationToken)
         {
-            var readStudentModel = _studentReadRepository.GetBySSN(request.SSN);
-            
-            _studentReadRepository.Delete(readStudentModel);
+            if (response.IsSuccess)
+            {
+                var readStudentModel = _studentReadRepository.GetBySSN(request.SSN);
+
+                _studentReadRepository.Delete(readStudentModel);
+            }
 
             return Task.CompletedTask;
         }
